Update PHP Mess Detector to 2.8 branch
This is still in beta (released some 19 hours ago) but it's been months
in the making and seems to run just fine. While not mentioned in the
changelog, a new 'UndefinedVariable' rule has been added but all the
errors it's picking up are in `exec()` calls, so we'll ignore it.

2.8 also adds an 'ansi' reported for coloured / more readable output in
the shell, so this switches both Travis and out Makefile to use it.