Update symfony/console to fix Travis bug
Since adding PHP 7.4, travis fails with an error in ArrayInput.php:
ErrorException: Trying to access array offset on value of type int

This was fixed in the 4.3 branch but initially composer update only took
us to the latest version of 4.2, presumably because of a conflict with
event-dispatcher. After first updating that, it took us straight to 4.4.