Add switch to enable creating a user from sites
Creates a new System User section in the site editor with a toggle
switch to say we want a user/group created when we save the site.

A toggle/label on its own looks a bit weird, so we wrap it in a segment
so it looks a bit more in line with the other fields on the page. The
font colour while the switch was enabled was broken in dark mode too, so
we also add a fix (requiring !important) in the darkmode overrides file.