Clean up FileEditor.js
This removes the `val` alias of the `value` parameter which seems
useless, and fixes the no-redeclare and no-cond-assign eslint errors.

The redeclare issue is just because we're using `var info` twice, so we
add it to the list of `var`s so that we're only assigning later.

The no-cond-assign rule expects `foo = bar` to be on its own line rather
than being used as a conditional in an if statement. It can be "ignored"
by wrapping it in brackets, but that would trigger another rule about
unnecessary brackets which we haven't (yet) enabled. Instead, we suck it
up and move the assignment outside of the condition while also renaming
it so that it makes more sense what it is vs guessing what 'm' could be.