Add more rules from --init with autodetection
After running eslint --init a second time to generate a new config with
code scanning enabled for the rule generation, the results were compared
against the docs and added to our config. Any "non-safe" rules that
would require changes are disabled or relaxed to avoid any errors.

There are still a couple in the docs that weren't in the generated
config, and likewise several in the config that don't seem to be in the
docs. For now, these are left out but can be added in later.