Enable creation of users with a home directory
When we select to create a user for a project, we'll probably want to
store the project's files somewhere specific to that particular user.

Our vagrant environment with Ubuntu doesn't enable setting/creation of
the home directory by default, so this updates the LinuxUser class to
support setting the -m/-M args to the useradd command, respectively
enabling/disabling the functionality.

Closes #201.