Show user details on project overview page
This adds a new column to the sites table so we can keep track of the
uid of a project's system user once it's successfully created. With
that, we're able to show user details in the overview including the
user's homedir, with a button to open that path in the file browser.