Fix bugs when a linked user doesn't exist
If a user doesn't exist and you try to open the location containing its
home directory in the file browser, it errors out because the user can't
be found so this defaults to showing '???' in those cases.

To avoid issues in the main sites list and anything using site models,
this also catches the `UserNotFoundException` and returns null when
there's no user defined or the defined user doesn't exist.