Update current data when saving sites
When you enable creation of a user and save the site, the switch stays
visible until you refresh because the site's user info is never updated.

To fix that issue we set the value of `current` (aliased to `tmpSite` in
the editor) to a new copy of the updated object from the sites list.

Probably should've done this sooner, but it doesn't really affect string
values since the user will have entered those prior to saving the site.