Support creating users with/without the user group
Adds a switch to the user create form that enables or disables the
creation and assignment of a group with the same name as the user,
adding either the -U or -N flag respectively to the useradd command.

If we're enabling the user group, we don't want to also pass a group
name or gid so we only show the group dropdown if we're in edit mode, or
the user group switch is disabled in create mode. In the latter instance
we default to group id 100 which on most systems is the 'users' group.