Don't require gid if user group is enabled
Previously the gid field was required regardless of whether we wanted to
set it or not, so a lot of places would set it to 0 which introduces
issues should we ever want to enable actually setting it to 0 (root),
but the chances of that happening are slim.

This adds a separate set of rules for creating vs updating a user so
that we can ignore the gid unless user_group is set to true. When we're
updating a user we make it nullable as it's not normally required.