Add support for changing a user's home directory
This replaces the plain text homedir with an input styled to fit inline
with the browse button so that the home directory can be modified.

To support it on the backend, a new setHomeDirectory method has been
added to handle the -d arg. Note that this will not create the new dir.

This implements part of #126.