Enable moving of the old home directory contents
Adds a switch below the homedir field if it's been modified which lets
you move the contents to the new home directory, which will be created
if it doesn't already exist. Switched on by default since it's likely
rare that anyone will need to rename a home directory without moving it.

Closes #126.