Add ability to set custom home directories
Adds an input to enter the path to a custom home directory location when
creating a new user, which can be combined with the switch to create it.

If we start to edit the home directory in the create form, it triggers
the edit version which was only checking whether tmpUser.dir was set.
That check was made to also consider edit mode to avoid UI switching.