Properly clone a group's users array to fix #175
No idea why this fixes it but somehow the sliced arrays weren't cloning
the data properly, so as soon as you added a user to the group in the
form, it would show up in the group's user list on the left as well.

Could just be the state.group vs group, but I cba trying it. This works.