Fix bug preventing all users' removal from a group
If you tried to remove all users from a group, or the only user if
there's only one, then it would silently fail and the user would still
be a member of said group. Since an empty array is falsey in value, we
fix this by simply checking if the members array is set.

We also clarify the 'something went wrong' errors since groups have two
commands that are called, so it's nice that we can spot the difference.