Refactor group listing and creation
First half (third?) of the big groups refactor which will finally deal
with the last bits of #141 once and for all. This is just the stuff
needed for the index and store methods so far, but anything more than
that would probably end up being a bit too much to digest in one commit.