Refactor validation and deletion of system groups
This adds the find method to System\User so that we can migrate the
delete logic into it, and moves the validation rules into a Request.

To keep the verboseness of the name field's errors, we also add three
new rule classes for comma, colon and whitespace/tab/newline chars.
Since the user names also have the same rules, we update that base
request too so we get proper errors and can remove ValidLinuxUser.