Finalise system groups refactor
This refactors the update method of the System\Groups controller, which
is the last of the refactor. The commit method has been split in two
here whereas the System\User class only has one. This is because saving
a group's users is a two-step process so the handling is very different.

Closes #141.