Add support for setting or changing a user's shell
This adds handling to the LinuxUser class for setting the -s option to
either useradd or usermod so we can set from create *and* update.

We default to zsh because, well... just because. Couldn't really list
all /bin/*sh matches in a dropdown for a few reasons. False positives
are many on my arch machine, and not all OSes use the same paths. Since
19.04 Ubuntu has switched to symlinking /bin to /usr/bin, but in 1804,
/usr/bin/bash doesn't exist and /usr/bin/zsh symlinks to /bin/zsh.

Closes #220.