Fix new groups not being visible until refresh
This is actually two issues combining to one common symptom. First is
that list returns gid/uid as strings whereas add returns an int, so when
the dropdown compares the new gid, string !== int. The other issue is
that of the new group not being known to the system as the list hasn't
been refreshed since the user was created with its group.

This fixes the former by simply converting any integer gids to a string,
and the latter by triggering the reload of group data when a user is
created, making sure to not unset the edit mode until it's done.

Fixes #217.