Fix issue with empty arg being added
Because the `-m` flag for move_home is only set if you want to enable
it, the result of the ternary is an empty string which gets set as the
arg and ends up breaking dirty checking. This is solved by simply
assigning the ternary first so that we can ensure it's not empty.