Test all untested system user/group bits
Does what it says on the tin. This removes the GROUP_UPDATE_FAILED case
from commitAdd because the only real way of testing a failure to write
to the /etc/group file would be to lock it. We can't do that with exec
because `flock` needs a command or file descriptor to be passed, and
running a command such as `sleep` or `cat` would prevent PHP from
continuing with execution. The only other alternative is the PHP
function `flock()`, but that can't write because it has no sudo.

In either case, we have no way of unlocking the file after.