Enable easy creation of system users (#219)
Adds a switch to the user editor (only visible in create mode) that will
enable passing the -r flag to assign a uid in the system user range, so
we no longer need to manually set a uid < 1000 to create a system user.

This is the first of two parts, groups will need a bit more work.