Fix phpstan errors about class return types
The self hint was indicating LinuxUser which we need to be able to call
other methods on it later, but as toggleArg was defined on the parent
LinuxCommand class, it was actually returning that instead. This moves
the functionality out into a trait so it can be injected directly into
the class that'll actually be calling the toggleArg method.

Also adds toArray as an abstract method on the parent so that stan can
stop complaining about calling a method that doesn't exist (but actually
does exist but only on the child classes).