Exclude file update test from travis runs
Travis seems unable to create the file even in a project dir, so we'll
just lump it into a group that we can ignore when we call phpunit from
the travis script.sh. Older PHP (7.3 iirc) was also complaining about
using string type hints on the class properties, so we ditch those too.

Since we're explicitly excluding certain tests by a group, we can
include the old free memory test for the stats bar as well. It still
doesn't pass in Travis so needs to be excluded there, but seems to be
just fine when tests are run locally in the vagrant vm.