Fix newlines being trimmed from files on update
This adds a selection of test for error handling of file updates and
fixes an issue along the way where you wouldn't be able to write
something like `"foo\n"` to a file without the `\n` being trimmed.

Switching InvalidArgumentException to ValidationException doesn't really
matter much, it's just easier to test things that way.