Handle 500 error when listing nonexistant paths
Adds handling for Symfony's `DirectoryNotFoundException` so that instead
of getting a 500 we can return a 404 that's then transformed into a
friendly error placeholder in the file manager. So that the user can
navigate back up to the previous directory (assuming it exists), we're
also including the `filepath` attribute in the error response.