Bring test coverage for FileManager to 100%
This adds some new tests to cover 403 and 415 (media/file type
unsupported) errors and tweaks some older bits that weren't tested.

The manual text/octal file permissions seem to be left behind from some
older code before it was updated to normalise the permission keys, so
this can likely be removed fairly safely without breaking anything.

In case of an error when opening a file, we now default to 418 rather
than falling back to it. Since the failing conditions for the fallback
error are inherently unknown, it's been difficult to test previously.
By defaulting instead, the same `$data` array is used by the by the
permission denied handling thus its test now also cover the default.

The 422 "failed loading file" fallback is also a relic from the past and
can safely be deleted. This was from a time when there wasn't any basic
mime/type checking on files, so image/font/binary files would trigger a
JSON error hence the try/catch when returning the response.