Add regression test for #214
While this issue was fixed in b18c8eba3a189b22a27329f8286a37e24790e633,
issue #214 was never referenced in the commit and a test wasn't added.

We could do this without using Arr::first and just dumping it all in the
assertJsonFragment call, but this way we get much cleaner failures.