Support creation of new files in the editor
If a file is not found when loading the editor, rather than showing the
404 error we'll now show the editor window with a blank slate.

To support saving when the file can't be loaded to start with, the save
function has been reworked to call a new create action, which will also
load the file and its path into state once creation succeeds. Since it's
effectively reloading the file, we set the loading state accordingly.