Handle deletion of directories
Previously the delete button was only visible for files, as folders
would trigger the unsupported filetype error. This now removes that
error if it's present after renaming a directory so the frontend can
continue resolving the promise that updates the list.

It's a bit crude, and repeats something similar in createDir, but it's
perfectly fine for what we need. Maybe refactor in future but no rush.