Add logs attribute to Site model
This adds `$site->logs` which will contain a key-based array of logs
based on the project type. For now we just add the PHP and Laravel logs,
but eventually this can be expanded to nginx access/error logs etc.

Since the case lines aren't recognised by phpdbg, this also updates the
vagrant environment and makefile to use php-pcov for coverage instead.