Add list of logs to Site details page
Adds a new section to the Site overviews that lists logs and their paths
in a table, opening the file editor when any row is clicked. The logs
weren't showing initially, so this also adds a test to ensure `logs` is
available as an attribute in the data returned from the Sites endpoint.