Ensure docroots are created with right permissions
Fixes an issue where document roots were being created with permissions
of 1341 because `755` was passed instead of the octal `0755`.

Interestingly this doesn't seem to affect docroots in test-skel, so the
test is set up to create its structure in /var/www instead.

Adding the third arg of `true` to mkdir also fixes an issue with Laravel
projects throwing an ErrorException as /foo/bar/public couldn't create
due to the parent not existing. Originally `basename()` was used so that
the parent directory would be created rather than the docroot itself,
but that interferes with non-Laravel projects that don't have `public/`.