Load logs directly to fix permission errors
Using the file browser to open logs doesn't work for PHP because it's
owned by root, which the web server doesn't have permission to access.

We workaround this by adding a site/:id/logs/:log endpoint that cats the
file via sudo to make sure we have every chance at reading it. This is
then called from the site detail page and injected into the log pane.