Reload first log when site is changed
While we had the mounted hook loading the first log, this wouldn't be
triggered when a second or third site was loaded from the sidebar, and
it also threw errors if a site didn't actually have logs.

To solve this we have a new initLog method that will unset the active
log if there aren't any to show. This is then called from `mounted`
instead of viewLog, as well as the beforeRouteUpdate hook to catch
internal page changes and reload logs when appropriate.

Oh, and it cleans up the Object.keys junk with a logNames helper too.