Fix relative log paths not loading
Relative paths should've had the document root prepended, but until now
that hasn't happened. Since we don't really have a *true* docroot for
Laravel projects yet, if there's a public dir at the end of the docroot
then we just trim that to get the actual project root.

Really need to add support for a true document root and public dir...