Add make target for building installer
In order to be easily usable on servers, we need to be able to fetch a
single script from somewhere and simply run it as-is. When running make
installer, first we merge the scripts making sure that install.sh goes
first since it's the one with the shebang.

When it's a single file we don't need the `source` lines so those are
removed along with the shellcheck comments that preceed them. We then
move the `main "$@"` line to the bottom of the file so that injected
functions can still be called without undefined errors.

`$$` here is actually `$`. We need to escape them or Make thinks we're
trying to reference a variable and breaks the sed patterns as a result.