Fix Laravel being unable to cache callback routes
When you try to cache the routes as part of `artisan optimize`, it fails
because we're using callbacks for the fallback and user profile routes.

This abstracts the (minimal) content from those callback functions into
a dedicated FallbackController for the api/frontend fallbacks, and a
single-action User\ShowProfile controller for the /api/user route.

Instead of using Laravel's `abort()` helper, we now throw the Symfony
exception ourselves so that the tests can correctly report coverage.