Rewrite edit_line function for better reliability
tl;dr: .env should now be much cleaner, and any existing values should
now be reliably handled to prevent them overriding the saved settings.

---

Various different values were causing issues with the previous code used
in the Vagrant bootstrap script, particularly involving "special" chars
used in passwords which would cause the sed patterns to break.

This completely reworks the function so that we first check for an exact
match and remove *all* other exact or empty matches, which both cleans
up duplicates in the .env and reduces load in the second loop.

If we don't find an exact match, the second loop will find the first
line that matches the option regardless of whether it has a value or
not. If it does we comment and append like, otherwise edit in-place as
we did previously. Once a value is set, all further matches will be
either commented or removed depending on whether a value is defined.

Along with a vastly different set of sed (and now awk) commands, we now
have two cleansing functions that escape values for the left (safe_srch)
or right (safe_repl) side of the sed given pattern. This should help to
avoid the majority of previous issues caused by values with symbols in.