Add 'servidor' system user
When running unattended (eg as a first-boot script on services like
DO/Vultr), Composer complains that both $HOME and $COMPOSER_HOME are
unset. While we could likely set the COMPOSER_HOME var to something
along the lines of /var/servidor/build/composer, we may as well have a
real system user with a home directory set to the install root.

This also updates all related commands to sudo via the servidor user,
and passes -H to sudo so that the home directory of servidor is used.