Add a dedicated field for laravel public folder
Until this point, Laravel's public folder was defined within the
document root, which has caused issues in the past with directory
creation and cloning projects. By separating the public dir into a
dedicated attribute, we can use the project_root when cloning,
regardless of whether it's a laravel or plain HTML project.

So that we can still easily get the full document_root, e.g. in the
nginx server block, we also add a new `getDocumentRootAttribute` method
on the model which replaces the original document_root attribute.