Fix dropdowns being wider than the site editor
For some reason when the form-field is only 4 columns wide, the dropdown
doesn't shrink down enough to fit within the container, so it ends up
going outside the viewport in some cases. Giving it one extra column of
breathing room seems to be enough to fix that, so we do.