Remove core.css from global Servidor layout
The core styles were previously added globally to fix issues with icons
in dark mode while also providing earlier styles as it was in Blade.
This however causes issues with duplicate style declarations as the dark
theme is essentially a full copy of Semantic UI.

This removes core.css from the global layout, instead using only it or
the dark styles - not both. As login elements are all inverted already,
that only ever needs the dark styles. In order to restore fonts, we also
import both the Google font and the Font Awesome icons from s-ui-sass.

There's a good chance this could adversely affect some auth templates
that were never ported to Vue, but I'm not even sure we're using 'em.