Inject CSS dynamically via Blade to fix hot reload
For styles to be reloaded correctly when using `npm run hot`, we should
really be using the `mix()` helper to load the CSS over the dev server,
but that's not possible within Vue as it relies on Blade functionality.

Here, we replace css/js paths with a `smart_asset` helper that calls one
of `mix()` or `asset()` based on environment. To access the paths in Vue
we set them in Blade, adding a new method in the darkmode plugin to get
the correct style set so we don't need to define them in every layout.