Fix light vs dark theme conflicts in `npm run hot`
When using hot reload, old styles seem to stick around causing issues
with navbar background colours among other things. To avoid conflicts,
this prefixes dark styles with a `.dark` class. While we could add the
class straight to `<div id="app">`, this adds a `themed-page` component
that generates the page using a `render` function to pass the current
theme as a class on the main `app` div.

Not only can we load styles directly in the component, eliminating the
need for `<link v-for...>` in layouts, but we can also generate the
wrapping tag dynamically which enables `<foo is="bar">` syntax in places
where more than just a `div` is needed.

For any 'replacement' component to receive the correct attributes, we
need to pass them along from the calling component VNode's `data`
object. Note that while the following will work as expected...
```html
<sui-grid is="themed-page" layout="login" textAlign="center">
// Result: <div class="ui center aligned grid dark">
```

...this will leave props in the markup (although classes are added):
```html
<themed-page is="sui-grid" layout="login" textAlign="center">
// Result: <div class="dark ui center aligned grid" layout="login">
```