Split vagrant provisioning into multiple steps
The project folder was originally configured to use 998 as the user and
primary group IDs in ceb733c1f0a5045ce1d2ff50c98104f3d05abf6f, where 998
is the ID of the servidor user/group... or was on my machine.

Turns out, that's not always the case--in #274 the servidor user got a
different UID/GID, resulting in a failed install due to permissions.

We can't set the owner/group to `servidor` until it exists, so this adds
an inline shell provisioner specifically for creating the user. To run
that before the installer though, we need a 3-step process. The first is
`vagrant up` because you can't `vagrant provision` before the VM is
created. Although `up` will fail to mount shares, we can then provision
the user before restarting the VM (to get shares back) to run install.

The process is long-winded and composer *still* has permission issues,
but this at least solves the race condition with the servidor user.