Fix composer install within Vagrant
Due to a bug in Virtualbox, there is a race condition when using vboxsf
for the synced folders where Composer extracts and deletes files faster
than Virtualbox can keep them in sync.

Adding the `--perfer-source` option to composer will stop it using the
compressed distribution packages to avoid the extract/delete issue.

Also adds a separate ./storage mount with the www-data group to fix
an issue with writing the Laravel error log.

**Related:**
* https://www.virtualbox.org/ticket/18776
* https://github.com/laravel/homestead/issues/1240#issuecomment-555951201