Run composer on host machine where available
Turns out running composer with --prefer-source is slooooooow. we're
talking like... 30 minutes slow. It's normally done in less than two.

This adds a check in the makefile to see if we have composer available,
and if so we'll run it before creating the Vagrant VM which will cut out
much of the process when it comes to composer install in the installer.

No need to make any changes in the installer itself as composer is smart
enough that it will only install any packages if they're different to
what was already installed 5 minutes previous. That's only likely to
happen if someone has a custom composer config that changes things.