Fix sed commands on OSX
The version of `sed` available on OSX is an outdated piece of garbage,
like most system utilities in the apple world. It doesn't like the
`{/.../,+1d}` expression to "find this line and delete it and the one
after", but since the lines in question are the ones that include the
install scripts, we kinda *need* to remove them.

Ironic that fixing this the easy way would create an Infinite Loop...

I'm not about to turn the sed command into something 20 lines long just
to satisfy OSX, so instead we check for gnu-sed before falling back to
sed-proper. If it turns out we're using `sed` and uname == Darwin, that
means they didn't brew install gnu-sed and thus we abort compilation.