Set file editor mode based on filename (#276)
Replaces `CodeMirror.findModeByExtension` with `findModeByFileName` to
detect nginx config files properly. If `CodeMirror.findModeByFileName`
doesn't match anything it automatically tries to find a match via
`findModeByExtension`. If that fails to return info as well, we try to
determine a mode by calling `CodeMirror.findModeByMIME`.

This is the discussion + pr in the CodeMirror repo:
* https://github.com/codemirror/CodeMirror/issues/2940
* https://github.com/codemirror/CodeMirror/pull/3027

Fixes #187