Add new create/list pages for projects
Beginnings of a complete reimagination of the whole sites/applications
business. The idea being the "project" is simply a named container that
holds one (or potentially more) application(s) added in a separate step.

Eventually databases will be a part of the projects, and a separate
backend/system section can be added for admins so in "project mode" we
won't need the sidebar or could use it to show the system stats instead.