Normalise buttons and headings
Adds next/cancel buttons in place of the randomly-labelled buttons and
adds a uniform heading to each step for a bit more context.

Eventually a back button would be better, but we'd only want to go back
so far - that is, once the template is set it shouldn't be changed.