Finalise the source location step
Moves the list of source providers into data so they can have a url
template set which allows us to replace the "user/repo" string for
supported providers and have the URL generated automatically.

If the selected host doesn't have a url format defined (ie it's the
custom option) then we just use the manually-set url field instead.

Also moves the 'code will be pulled from' paragraph up one spot so that
its order matches that shown in the steps.