Abstract source location step to own component
This was a little more involved than the rest because it uses temporary
fields to decide the actual repository url that'll be returned by it.

To save messing about, we simply pass the original repo name and
provider along with the rest of the data, then save it to a new
`extraData` object that's passed to the ConfirmationText view.