Prompt before leaving New Project page
To avoid losing changes, when the back button is clicked we should
really show a confirmation, not just when a user clicks cancel.

This converts the existing cancel modal to a component loaded in the
main NewProject.vue page so that we can trigger it both from the page's
beforeRouteLeave handler and also within forms in child components.

To avoid the leave handler triggering the prompt a second time if the
'discard and go back' button is clicked, we have to save a bypass bool
in the state. It would be nice to pass a temporary option with the route
object, but that doesn't appear to be possible.