Generate source repo URIs on the backend
This moves the generation of source URIs into the backend such that when
creating a project, we need only supply the 'user/repository' format as
long as the selected provider has a corresponding uri in the const.

Some duplication is avoided since we're still converting the user/repo
string manually into the correct clone url in the create form, but
eventually providers could be provided (heh) entirely by the backend.