Update project list and start on the viewer
Adds a new ListProjects controller to show real projects in the list
loaded via vuex so that we can get to work on the project viewer.

For now this only copies over the Source Files segment (minus the pull
latest code button) which is in itself likely to change a fair bit.

Project Root is missing as it needs to be generated in the model.