Move log-reading logic into new LogFile class
LogFile takes the project's `Application` instance along with a title
and path. Rather than calling `readLog('php')` on the app model, we can
now simply cast it with `(string) $log`.

The logs list is also updated to return a simple list of `name => title`
objects, which avoids exposing full paths through the API unnecessarily.

Note app-specific logs such as the Laravel log are still not working
currently as they rely on the system user and base path being ported.