Add system user status+creation to project viewer
This ports the system user functionality from Sites, utilising a model
observer in place of much of the controller code used previously. Since
we're no longer generating the username in the frontend, we also negate
the need for saving it in a `system_user` column. Currently a project
can't be renamed so this is not an issue and can be accounted for later.

When a project is created, it's apps are now added with `saveMany()`
instead of `createMany()` to ensure the created/saved events are fired.
So that users are only created where necessary, we track whether to
create a system user via the new `requiresUser():bool` template method.