Fix failing Project Mismatch exception test
Fixes an issue where being in production mode caused the "Project
mismatch" exception message to be gobbled up by Laravel's exception
handler. More of a temporary solution, but this disables exception
handling for the duration of the request being tested, and rather than
checking the response we tell phpunit that we're expecting an Exception.