Port list filtering and selection to Projects
Adds a search box to the top-right of the projects page, which can be
used to filter the list of projects based on their names.

Unlike the functionality in Sites, this will not create a project. What
it will do, however, is view not only an exact match, but also the first
filtered result when no exact match can be found in available projects.

Also adds a min-width to the domain column, since that seems to shrink a
bit too much when the table's rows are filtered down to shorter entries.