Add support for domain redirects via archive.org
Enables the archive "template" and adds all the necessary front/backend
code for creating a project with redirects instead of an application.
Rather than supporting any arbitrary source/destination, for now the
frontend is limited to full-domain redirects. These use a generated URL
to the WayBack Machine based on a given domain and archive date/time.

Methods to enable or disable the nginx config for an app have been moved
to a new TogglesNginxConfigs trait, also used by Projects\Redirect.