Prevent toggling nginx config without domain name
An application doesn't necessarily have a domain, but previously the app
observer would try to create and enable/disable the config regardless.

This adds a Template interface so that we can check the class instance
in the TogglesNginxConfigs trait whether it's Html, Laravel or whatever.
Since psalm was complaining about types without a @var declaration, we
also add a `getApp()` method so its type can be made explicit.