Update laravel/passport to 9.4, lowest 9.x w/ php8
PHP 8 support was only added to the 9.x branch, so this bumps to the
next major version. We should really add a migration[1] that adds a
provider column, but given the only installs are dev copies that can
simply be reinstalled, and travis does a fresh install anyway, we're
probably able to get away without it. Famous last words and all that.

[1]: https://github.com/laravel/passport/blob/v9.4.0/UPGRADE.md#support-for-multiple-guards