Update to phpunit 9 and arraysubset-asserts 0.2.1
PHP 8 support is added in 0.2.1 for arraysubset-asserts. I considered
requiring `^0.1.0 || ^0.2.1`, but since the latter requires phpunit^9.0,
we'd end up needing to have both versions 8.5 and 9.x of phpunit too.

The oldest phpunit with support for PHP 8 is 9.3.0 which we need to
specify otherwise we get 9.1.5 for whatever reason. Since we've gotta be
more verbose about it, we may as well go straight for updating to 9.5.