Update php-cs-fixer to 2.18.2
This updates cs-fixer and bumps the migration rulesets to 8.0. The 7.4
set includes `use_arrow_functions`, affecting 3 or 4 files. The arrow
function in FileManagerTest was a tad long, so that test is reworked.

Also adds the risky PHPUnit84Migration ruleset, which flagged a single
case of `assertEquals()` being used instead of `assertCount()`.

Since new rules apparently get added in minor versions which can cause
things to break unexpectedly, this uses `~` instead of `^` for the
version requirement so that we only get patch releases for cs-fixer.

2.17.0 also started to include `php_unit_method_casing` within the
Symfony ruleset, so this disables it. I would've went with snake_case,
but there are a small number of tests that mix both snake and camel
casing where they test a specific method, eg test_setFoo_does_a_thing.