Restore guzzlehttp/guzzle and update both it+psalm
Guzzle was previously required by Passport, but apparently when I used
it in the Login code, it was never added as an explicit requirement.

Passport has since removed the guzzle dependency, so we take this
opportunity to also update it to 7.x (passport required 6.x previously).

This also removes the ocramius/package-versions lock which was
previously required for 7.3 to pass in travis due to older psalm
dependencies. In theory, this should clean up the problem with
composer/package-versions-deprecated being locked to 1.8.0 (php^7).

Use my fork (for now) for phpmnd with php8 support