Restore (and update) guzzlehttp/guzzle
It was previously required by Passport, but apparently when I used it in
the Login code, I never added it to composer an explicit requirement.

Passport has since removed the guzzle dependency, so we take this
opportunity to also update it to 7.x (passport required 6.x previously).

Thanks to psalm for flagging the missing GuzzleHttp\Client class, and to
phpstan for pointing out the getBody() ternary is unnecessary now.