Move PHP linting / static analysis to its own job
Removes the need for CI_*_VERSION entirely as all CS/SA tools are now
run via make, directly from the workflow in a dedicated lint-php job.

To support that, php-cs-fixer and phpmnd needed a way to set extra
flags, so make will now accept CS_ARGS and MND_ARGS with the desired
options and arguments to be passed along to the respective command.

Also removes RUN_MODE, which I thought I stripped already.