Fix composer in lint job, update actions/cache
Modifies the lint-php job to use the composer cache, and updates the
cache action to v2 in both the php jobs as well as the js one.

When creating cache keys, the os var has been dropped as only one is
used for node, and the difference in ubuntu version isn't likely to
affect any composer packages that get installed. For fallbacks, we set
restore-keys to "composer-[phpver]-". Any less strict than that and we
would have issues with php 7.4 trying to use packages built for 8.x.

The composer install step has also been moved into the main workflow
file, with nick-invision/retry@v2 used in place of the old travis_retry
script, now removed. We also make use of this now for npm clean-install.

We're caching .cache/composer now, not vendor/, so we should probably
run the actual install using that cache rather than leaving the other
steps wondering where all their php tools disappeared to after 1st run.