Fix test that checks system root (/) listing
On Ubuntu 18.04, both /bin and the /lib* dirs are all physical
directories, but that changes on 20.04 where they're symlinks to /usr/*.

Permissions on the targets remain at 755, but as symlink permissions are
all 777 it's breaking the tests on Ubuntu 20.04. This drops the perms
array from the test because it's surplus to what we're actually testing.