Only run GHA on PR, *or* push to master/dev
It'd be nice to run on pushes to a branch that is not used for a pull
request, but it seems there's currently no way to do that in a workflow.

Running on push *and* pull_request for *all* branches, we end up with
duplicate workflow runs when a PR is created from a non-fork branch.

This fixes it by restricting push runs to master and develop **only**,
with the tradeoff that any local branches will need to be triggered
manually if we want to run CI pre-PR. Small price to pay, though.