Add Laravel plugin for Psalm
Enables the Laravel plugin which essentially just pipes ide helper into
Psalm so it can get a better idea of what's going on. In doing so, it
also revealed another Response|JsonResponse that wasn't caught
previously, so that's fixed here.

Since psalm has issues in the TogglesNginxConfigs trait where it would
try to call getApp() even when $this isn't a Template instance, we also
add a new Domainable interface used by both the base Html template *and*
the Redirect model, allowing us to simply call $this->domainName() on
anything that implements it.